# kunk-api/src/routes/authAssociate.js
'use strict';

const { Router } = require('express');
const associateAuthRepository = require('../repositories/associateAuthRepository');
const { requireAssociate } = require('../middleware/requireAssociate');
const { ok } = require('../utils/response');
const { env } = require('../config/env');

const router = Router();

function setAssociateCookie(res, sessionToken, expires) {
  res.cookie('associate_session', sessionToken, {
    httpOnly: true,
    secure: env.cookieSecure,
    sameSite: 'lax',
    path: '/',
    maxAge: env.sessionMaxHours * 3600 * 1000,
    expires,
  });
}

function clearAssociateCookie(res) {
  res.cookie('associate_session', '', {
    httpOnly: true,
    secure: env.cookieSecure,
    sameSite: 'lax',
    path: '/',
    maxAge: 0,
  });
}

router.post('/register-email', async (req, res, next) => {
  try {
    const { email, password } = req.body || {};
    const { user, sessionToken, expires } = await associateAuthRepository.registerEmail(email, password);
    setAssociateCookie(res, sessionToken, expires);
    res.status(201).json(ok({ user }));
  } catch (err) {
    next(err);
  }
});

router.post('/login', async (req, res, next) => {
  try {
    const { email, password } = req.body || {};
    const { user, sessionToken, expires } = await associateAuthRepository.login(email, password);
    setAssociateCookie(res, sessionToken, expires);
    res.json(ok({ user }));
  } catch (err) {
    next(err);
  }
});

router.post('/logout', async (req, res, next) => {
  try {
    await associateAuthRepository.logout(req.cookies?.associate_session);
    clearAssociateCookie(res);
    res.json(ok({ ok: true }));
  } catch (err) {
    next(err);
  }
});

router.get('/me', requireAssociate, async (req, res) => {
  res.json(ok({ user: req.associate }));
});

router.post('/forgot-password', async (req, res, next) => {
  try {
    const data = await associateAuthRepository.forgotPassword(req.body?.email);
    res.json(ok(data));
  } catch (err) {
    next(err);
  }
});

router.post('/reset-password', async (req, res, next) => {
  try {
    const { token, password } = req.body || {};
    const data = await associateAuthRepository.resetPassword(token, password);
    res.json(ok(data));
  } catch (err) {
    next(err);
  }
});

module.exports = router;